Other algorithms comparison/CWR/requirements.txt
numpy==1.16.2
pandas==0.24.2
torch==1.1.0
torchvision==0.2.2
Pillow==6.0.0
